+++ crates/stackable-shared/src/lib.rs
@@ -2,6 +2,7 @@
 //! workspace.
 
 pub mod crd;
+pub mod health;
 pub mod secret;
 pub mod time;
 pub mod yaml;