+++ crates/stackable-webhook/CHANGELOG.md
@@ -7,9 +7,9 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - BREAKING: The `WebhookServer` now serves a `/ready` endpoint for a startup probe.
-  For that, `WebhookServer::new` takes an additional `HealthCheckRegistry` argument ([#]).
+  For that, `WebhookServer::new` takes an additional `HealthCheckRegistry` argument ([#1272]).
 
-[#]
+[#1272]: https://github.com/stackabletech/operator-rs/pull/1272
 
 ## [0.9.2] - 2026-07-06