+++ crates/stackable-webhook/CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.9.3] - 2026-09-23
+
 ### Added
 
 - Add `health` module containing `HealthCheck` and `HealthCheckRegistry` for health endpoints ([#1272]).